+++ gdsfactory/components/coupler_full.py
@@ -6,7 +6,7 @@ import picwriter.components as pc
 import gdsfactory as gf
 from gdsfactory.component import Component
 from gdsfactory.components.waveguide_template import strip
-from gdsfactory.types import ComponentFactory
+from gdsfactory.types import ComponentSpec
 
 
 @gf.cell
@@ -18,7 +18,7 @@ def coupler_full(
     parity: int = 1,
     port: Tuple[int, int] = (0, 0),
     direction: str = "EAST",
-    waveguide_template: ComponentFactory = strip,
+    waveguide_template: ComponentSpec = strip,
     **kwargs
 ) -> Component:
     """Adiabatic Full Coupler.